requirements.txt
Django==4.0.4
django-mathfilters==1.0.0
python-dateutil==2.8.2
numpy==1.22.3
timezonefinder==5.2.0
pytz==2022.1
requests==2.27.1
xmltodict==0.12.0